scrapy/requirements.txt
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.41
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
mysql-connector-python==9.1.0
packaging==24.2
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.3.0
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==69.0.3
smmap==5.0.1
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
zope.interface==7.2
